Alle skjema er lagt inn både for kort og detaljside.
Grafik flyttet øverst i hvert skjema.
Skilt ut endre-knapp i egen komponent for gjenbruk begge steder.
