+++ src/app/components/observation/registration-edit-button/registration-edit-button.component.html
@@ -1,0 +1,9 @@
+@if (userCanEdit()) {
+  <ion-button fill="clear" class="edit-button" (click)="edit()" [disabled]="isLoadingObsForEdit()">
+    <ion-icon name="create-outline"></ion-icon>
+    {{ "REGISTRATION.EDIT" | translate }}
+    @if (isLoadingObsForEdit()) {
+      <ion-spinner name="dots" slot="end"></ion-spinner>
+    }
+  </ion-button>
+}
