+++ src/app/components/observation/registrations/avalanche-activity-view/avalanche-activity-view.component.html
@@ -3,6 +3,19 @@
 @if (title()) {
   <h4 class="title">{{ title() }}</h4>
 }
+<div class="exposition-graphics">
+  @if (exposition()) {
+    <app-exposition [value]="exposition()" text="test:)" />
+  }
+  @if (exposedHeightComboTid()) {
+    <app-exposed-height
+      [exposedHeightComboTid]="exposedHeightComboTid()"
+      [exposedHeight1]="exposedHeight1()"
+      [exposedHeight2]="exposedHeight2()"
+    />
+  }
+</div>
+
 @if (comment()) {
   <ion-text color="dark">{{ comment() }}</ion-text>
 }
@@ -30,15 +43,3 @@
 @if (propagation()) {
   <app-key-value [key]="'REGISTRATION.SNOW.AVALANCHE_PROBLEM.PROPAGATION' | translate" [value]="propagation()" />
 }
-<div class="exposition-graphics">
-  @if (exposition()) {
-    <app-exposition [value]="exposition()" text="test:)" />
-  }
-  @if (exposedHeightComboTid()) {
-    <app-exposed-height
-      [exposedHeightComboTid]="exposedHeightComboTid()"
-      [exposedHeight1]="exposedHeight1()"
-      [exposedHeight2]="exposedHeight2()"
-    />
-  }
-</div>
