+++ src/app/components/observation/registrations/ice-thickness-view/ice-thickness-view.component.html
@@ -1,3 +1,3 @@
 <!-- Visning av istykkelse -->
-<app-summary [summaries]="summaries()"></app-summary>
 <app-ice-plot [url]="plotUrl()"></app-ice-plot>
+<app-summary [summaries]="summaries()"></app-summary>
