+++ src/app/modules/registration/edit-registration-helper-functions.ts
@@ -1,4 +1,7 @@
+import { WritableSignal } from '@angular/core';
+import { Observable } from 'rxjs';
 import { RegistrationViewModel, MyPageData } from 'src/app/modules/common-regobs-api/models';
+import { RegistrationService } from '../common-regobs-api';
 
 export type EditMode = 'EDIT_AS_MODERATOR' | 'EDIT_OWN_REGISTRATION';
 
@@ -6,7 +9,7 @@ export function isSameObserver(reg: RegistrationViewModel, observer: MyPageData)
   if (!observer) {
     return false;
   }
-  return observer.NickName === reg.Observer.NickName; // TODO: Change to ObserverID, when implemented in API model (MyPageData)
+  return observer.ObserverId === reg.Observer.ObserverID;
 }
 
 export function isInGroup(reg: RegistrationViewModel, observer: MyPageData): boolean {
@@ -37,4 +40,26 @@ export function checkEditPriviliges(reg: RegistrationViewModel, observer: MyPage
     return 'EDIT_OWN_REGISTRATION';
   }
   return undefined;
+}
+
+function isRegistrationOlderThan2days(reg: RegistrationViewModel): boolean {
+  const now = new Date();
+  const twoDaysAgo = new Date(now.getTime() - 2 * 24 * 60 * 60 * 1000);
+  const registrationDate = new Date(reg?.DtRegTime);
+  return registrationDate && registrationDate < twoDaysAgo;
+}
+
+/**
+ * TRUE hvis observatør har lov til å endre angitt observasjon.
+ * En "vanlig" observatør kan kun endre egne observasjoner som er yngre enn 2 dager.
+ * Moderator kan endre alle observasjoner uansett.
+ */
+export function canEditRegistration(reg?: RegistrationViewModel, observer?: MyPageData): boolean {
+  if (!observer || !reg) {
+    return false;
+  }
+  const editMode = checkEditPriviliges(reg, observer);
+  return (
+    (editMode === 'EDIT_OWN_REGISTRATION' && !isRegistrationOlderThan2days(reg)) || editMode === 'EDIT_AS_MODERATOR'
+  );
 }
