+++ src/assets/i18n/nb.json
@@ -429,6 +429,10 @@
         "MESSAGE": "Vil du slette denne observasjonen fra Regobs for godt?"
       }
     },
+    "DETAILS": {
+      "NOT_FOUND": "Fant ingen observasjon med denne ID'en",
+      "GONE": "Denne observasjonen er slettet."
+    },
     "DIRT": {
       "LAND_SLIDE_OBS": {
         "AREA_TOTAL": "Tegn hele skredarealet",
@@ -670,7 +674,7 @@
         "TITLE": "Skredhendelse",
         "TRAFFIC_OBSTRUCTED": "Trafikk hindret",
         "TRAJECTORY_NAME": "Skredbanenavn",
-        "VALID_EXPOSITION": "Eksposisjon"
+        "VALID_EXPOSITION": "Himmelretning"
       },
       "AVALANCHE_PROBLEM": {
         "ADD_NEW": "Legg til nytt skredproblem",
@@ -691,7 +695,7 @@
         "TITLE": "Skredproblem",
         "TOP": "Øverst",
         "UNKNOWN_TYPE": "Ukjent svakt lag",
-        "VALID_EXPOSITION": "Eksposisjon",
+        "VALID_EXPOSITION": "Himmelretning",
         "WEAK_LAYER": "Svakt lag",
         "WEAK_LAYER_DEPTH": "Svakt lags dybde"
       },
@@ -788,12 +792,12 @@
         "SUBFORMS": "Underskjema",
         "SUPERIOR": "Overordnet",
         "TITLE": "Snøprofil",
-        "VALID_EXPOSITION": "Eksposisjon",
+        "VALID_EXPOSITION": "Himmelretning",
         "WEST": "V"
       },
       "SNOW_SURFACE": {
         "COMMENT_HEADER": "Beskriv snødekket",
-        "COMMENT_PLACEHOLDER": "Hvordan er snøfordelingen generelt ifht. høyde og eksposisjon. Er det forskjell i regionen?",
+        "COMMENT_PLACEHOLDER": "Hvordan er snøfordelingen generelt mtp. høyde og himmelretning? Er det forskjell i regionen?",
         "HEIGHT_LIMIT_LAYERED_SNOW": "Øvre grense for snø uten lagdeling (moh)",
         "HIGHT_LIMIT_LAYERED_SNOW_PLACEHOLDER": "For snø uten farlig lagdeling",
         "NEW_SNOW_DEPTH": "Nysnø siste 24t (cm)",
